Refactor code in CartoBinaryTileLayer
